refactor(splits ownable): rename to splits balances
