+++ x/splits/queries/balances/query_response.go
@@ -1,7 +1,7 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package ownable
+package balances
 
 import (
 	sdkTypes "github.com/cosmos/cosmos-sdk/types"
